Rename `news` package to `getnews`
